requirements.txt
GitPython==3.1.30
nltk==3.8.1
PyDriller==2.4.1
PyGithub==1.58.0
requests==2.28.1
selenium==4.8.2
strsimpy==0.2.1
